Move methods into AttachmentPicker component
